feat(cli): model-strategy, slice-duration, isl-block-size, fixed-sched offsets, sketch
- --model-selection-strategy -> models.strategy.
- --slice-duration -> metrics.slice_duration_seconds.
- --synthetic-input-tokens-block-size/--isl-block-size -> prompts.block_size.
- --fixed-schedule-start/end-offset -> i64 phase offsets (auto_offset=false
  when offsets set).
- --sketch-metrics -> metrics.sketch + drops per-record records_path (oracle
  now sets the env like profile.py).
- config_extra/sched_offset/sketch goldens; 33 flag fixtures green.

+++ rust/cli/src/model/phase.rs
@@ -124,10 +124,10 @@ pub enum PhaseKind {
         auto_offset: bool,
         /// Optional explicit start offset.
         #[serde(default, skip_serializing_if = "Option::is_none")]
-        start_offset: Option<f64>,
+        start_offset: Option<i64>,
         /// Optional explicit end offset.
         #[serde(default, skip_serializing_if = "Option::is_none")]
-        end_offset: Option<f64>,
+        end_offset: Option<i64>,
     },
 }
